+++ tests/Tests/Exception/ChainExceptionTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Exchanger\Tests\Exception;
 
 use Exchanger\Exception\ChainException;
-use Exchanger\Exception\InternalException;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -39,7 +38,7 @@ class ChainExceptionTest extends TestCase
 
     public function setUp()
     {
-        $this->exception1 = new InternalException('Something bad happened.');
+        $this->exception1 = new \Exception('Something bad happened.');
         $this->exception2 = new \Exception('General exception.');
 
         $this->chainException = new ChainException([
@@ -53,7 +52,7 @@ class ChainExceptionTest extends TestCase
      */
     public function it_should_have_a_descriptive_error_message()
     {
-        $this->assertSame("The chain resulted in 2 exception(s):\r\nExchanger\\Exception\\InternalException: Something bad happened.\r\nException: General exception.", $this->chainException->getMessage());
+        $this->assertSame("The chain resulted in 2 exception(s):\r\nException: Something bad happened.\r\nException: General exception.", $this->chainException->getMessage());
         $this->assertCount(2, $this->chainException->getExceptions());
     }
 }